+++ Project/src/main/java/com/asgdrones/drones/services/InstructorServiceInterface.java
@@ -3,5 +3,5 @@ package com.asgdrones.drones.services;
 import java.util.List;
 
 public interface InstructorServiceInterface {
-    public List<String> getInstructorAddresses(Long loginID);
+    public List<String> getInstructorAddresses(Integer loginID);
 }
